ajout partiel function desactication trainings

+++ src/main/java/fr/fms/SpringApiRest/web/TrainingController.java
@@ -1,5 +1,6 @@
 package fr.fms.SpringApiRest.web;
 
+import fr.fms.SpringApiRest.dao.TrainingRepository;
 import fr.fms.SpringApiRest.entities.Training;
 import fr.fms.SpringApiRest.exception.RecordNotFoundException;
 import fr.fms.SpringApiRest.service.ImplTrainingService;
@@ -20,21 +21,28 @@ public class TrainingController {
     @Autowired
     private ImplTrainingService implTrainingService;
 
+    @Autowired
+    private TrainingRepository trainingRepository;
+
     @GetMapping("/trainings")
-    public List<Training> allTrainings(){
+    public List<Training> allTrainings()
+    {
         List <Training> lT = implTrainingService.getTrainings();
         return implTrainingService.getTrainings();
     }
 
     @GetMapping("/trainings/category/{id}")
-    public List<Training> trainingsByCategory(@PathVariable("id") Long id){
+    public List<Training> trainingsByCategory(@PathVariable("id") Long id)
+    {
         return implTrainingService.getTrainingsByCategory(id);
     }
 
     @PostMapping("/trainings")
-    public ResponseEntity<Training> saveTraining(@RequestBody Training t){
+    public ResponseEntity<Training> saveTraining(@RequestBody Training t)
+    {
         Training training = implTrainingService.saveTraining(t);
-        if(Objects.isNull(training)){
+        if(Objects.isNull(training))
+        {
             return ResponseEntity.noContent().build();
         }
         URI location = ServletUriComponentsBuilder
@@ -46,9 +54,10 @@ public class TrainingController {
     }
 
     @PostMapping("/update/{id}")
-    public ResponseEntity<Training> updateTraining(@PathVariable("id") Long id , @RequestBody Training t){
+    public ResponseEntity<Training> updateTraining(@PathVariable("id") Long id , @RequestBody Training t)
+    {
         Optional<Training> training = implTrainingService.readTraining(id);
-        System.out.println(training);
+        System.out.println("training " + training);
         if(training.isPresent())
         {
             Training train = training.get();
@@ -57,6 +66,7 @@ public class TrainingController {
             train.setPrice(t.getPrice());
             train.setCategory(t.getCategory());
             implTrainingService.saveTraining(train);
+            System.out.println("train " + train);
             return ResponseEntity.ok(train); // Retourne le Training mis à jour avec un statut 200 OK
         } else {
             return ResponseEntity.notFound().build(); // Retourne un statut 404 Not Found si l'entité n'est pas trouvée
@@ -64,17 +74,59 @@ public class TrainingController {
     }
 
     @DeleteMapping("/trainings/{id}")
-    public ResponseEntity<Void>deleteTraining(@PathVariable("id") Long id){
-        implTrainingService.deleteTraining(id);
+    public ResponseEntity<Void>deleteTraining(@PathVariable("id") Long id)
+    {
+        Optional<Training> getTrain = implTrainingService.readTraining(id);
+        if(getTrain.isPresent())
+        {
+            if(getTrain.get().isOnOrder()) {
+                Training train = getTrain.get();
+                train.setActive(false);
+                implTrainingService.saveTraining(train);
+            }
+            else
+            {
+                implTrainingService.deleteTraining(id);
+            }
+        }
         return ResponseEntity.noContent().build();
     }
 
     @GetMapping("/trainings/{id}")
-    public Training getTrainingById(@PathVariable("id") Long id){
-
+    public Training getTrainingById(@PathVariable("id") Long id)
+    {
         return implTrainingService.readTraining(id)
                 .orElseThrow(() -> new RecordNotFoundException("Id de formation " + id + " n'existe pas"));
     }
 
+    @PostMapping("/updateOnOrder")
+    public ResponseEntity<Training> updateOnOrder(@RequestBody Long t)
+    {
+        Optional<Training> training = implTrainingService.readTraining(t);
+        System.out.println("training " + training);
+        if(training.isPresent())
+        {
+            Training train = training.get();
+            train.setOnOrder(true);
+            implTrainingService.saveTraining(train);
+            System.out.println("train " + train);
+            return ResponseEntity.ok(train); // Retourne le Training mis à jour avec un statut 200 OK
+        } else {
+            return ResponseEntity.notFound().build(); // Retourne un statut 404 Not Found si l'entité n'est pas trouvée
+        }
+    }
+
+    @GetMapping("/trainingsActive")
+    public List<Training> allTrainingsActive()
+    {
+        List <Training> lT = implTrainingService.getByActive();
+        return implTrainingService.getByActive();
+    }
+
+    @GetMapping("/trainings/categoryActive/{id}")
+    public List<Training> trainingsByCategoryActive(@PathVariable("id") Long id)
+    {
+        return implTrainingService.getTrainingsByCategoryAndActive(id);
+    }
 
 }
